update product Update & Delete Auth in productController

+++ backend/controllers/productController.js
@@ -50,6 +50,12 @@ exports.updateProduct = catchAsyncErr(async(req,res,next)=>{
     if(!product) {
         return next(new ErrorHandler(404,"product Not Found"));
     }
+    //Checking user is admin or seller of the product
+    const isProductOwner = product.populate('user')
+      .then((u) => u.user.toString() === req.user._id.toString());
+      if(!isProductOwner || req.user.role !== "admin"){
+        return next(new ErrorHandler(401,"You are not authorized to Update this product"));
+      }  
     product =  await Product.findByIdAndUpdate(req.params.id,req.body,{
         new:true,
         runValidators:true,
@@ -62,14 +68,21 @@ exports.updateProduct = catchAsyncErr(async(req,res,next)=>{
 
 });
 
-//Controller for delete a product --Admin
+//Controller for delete a product --Admin --Seller
 exports.deleteProduct = catchAsyncErr(async(req,res,next)=>{
     let product = await Product.findById(req.params.id);
 
     if(!product){
-        return next(new ErrorHandler(404,"product Not Found"));
+      return next(new ErrorHandler(404,"product Not Found"));
     }
-    // await product.remove();
+    //Checking user is admin or seller of the product
+    const isProductOwner = product.populate('user')
+      .then((u) => u.user.toString() === req.user._id.toString());
+
+    if(!isProductOwner || req.user.role !== "admin"){
+      return next(new ErrorHandler(401,"You are not authorized to delete this product"));
+    }
+    // await Product.findOneAndDelete({ _id: req.params.id });
      product = await Product.findByIdAndRemove(req.params.id,req.body);
     res.status(200).json({
         success:true,
